refactor: implement route for admin dashboard
