+++ Muzik/src/containers/ContentContainer.jsx
@@ -1,12 +1,17 @@
 import SongsContainer from "./SongsContainer";
-import Admin from "../pages/Admin";
 import { getUserFavorites, saveSong } from "../api";
 import { useState, useEffect } from "react";
 import MapPage from "./Map";
+import { Notify } from "../utils/toast";
+import { Spinner } from "react-spinner-toolkit";
 export default function ContentContainer({ currentSection, userLat, userLng }) {
   const [favorites, setFavorites] = useState([]);
+  const [isLoaded, setIsLoaded] = useState(false);
   useEffect(() => {
-    getUserFavorites().then((data) => setFavorites(data.results));
+    getUserFavorites()
+      .then((data) => setFavorites(data.results))
+      .catch((error) => Notify())
+      .finally(() => setIsLoaded(true));
   }, []);
   const handleAddToFavorite = async (song, lat, lng) => {
     await saveSong(
@@ -22,6 +27,20 @@ export default function ContentContainer({ currentSection, userLat, userLng }) {
       }
     });
   };
+  if (!isLoaded) {
+    return (
+      <div className="loading-container">
+        <Spinner
+          shape="threeDots"
+          color="#888"
+          loading
+          speed={1}
+          size={300}
+          transition={true}
+        />
+      </div>
+    );
+  }
   return (
     <>
       {currentSection === "home" && (
@@ -40,7 +59,6 @@ export default function ContentContainer({ currentSection, userLat, userLng }) {
           favorites={favorites}
         />
       )}
-      {currentSection === "admin" && <Admin />}
     </>
   );
 }
